refactor: Remove message field from Generate_Snake.yml
Refactored the Generate_Snake.yml file to remove the message field. The message field was previously used for specifying the commit message when pushing the generated snake effect files to GitHub. However, it is no longer needed as the commit message can be automatically generated based on the workflow action.