docs: Comment out env field in Generate_Snake.yml
Added comments to the env field in the Generate_Snake.yml file to provide clarification and explanation. The env field was previously used for defining environment variables, but it is currently not being used in the workflow. The comments help to document the purpose of the field and its potential use in the future.