feat: Add _config.yml and set theme
Added the _config.yml file to the project, which contains the configuration settings for the site. Set the theme option in the _config.yml file to specify the desired theme for the site.

The addition of the _config.yml file allows for centralized management of site configuration settings and provides flexibility to customize various aspects of the site, including the theme.